pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "Ashes.live"
description = "API backend for a fan-developed deckbuilder and community website for the card game Ashes Reborn"
version = "2.0.0a1"
license = "ISC"
readme = "README.md"
authors = ["Ian Beck <ian@onecrayon.com>"]
packages =[
    { include = "api" }
]

[tool.poetry.dependencies]
python = "^3.11.9"
fastapi = "^0.116.1"
gunicorn = "^23.0.0"
sqlalchemy = "~1.3.23"
alembic = "^1.4.2"
psycopg2 = "^2.8.5"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
email-validator = "^2.2.0"
# Install uvicorn with the pieces from uvicorn[standard] that we will actually use
uvicorn = "^0.35.0"
uvloop = "^0.21.0"
httptools = "^0.6.1"
pydantic-settings = "^2.4.0"
httpx = "^0.28.1"
sendgrid = "^6.12.4"

[tool.poetry.group.dev.dependencies]
# These are used by the test suite
pytest = "^8.3.2"
pytest-cov = "^6.0.0"
sqlalchemy-utils = "^0.41.1"
freezegun = "^1.1.0"
# These are used by `make format`
black = "^25.1.0"
isort = "^6.0.1"

[tool.poetry.group.vscode.dependencies]
# These are used by VSCode
pylint = "^3.2.6"
rope = "^1.2.0"

[tool.isort]
profile = "black"
multi_line_output = 3
